api/requirements.txt
numpy
pandas
requests
psycopg2-binary==2.9.2
flask
sklearn
